Plug JWT Verifier logic into Credentials responses
