+++ auth0/src/main/java/com/auth0/android/authentication/AuthenticationAPIClient.java
@@ -47,8 +47,6 @@ import com.auth0.android.result.DatabaseUser;
 import com.auth0.android.result.Delegation;
 import com.auth0.android.result.UserProfile;
 import com.auth0.android.util.Telemetry;
-import com.auth0.android.verification.JwkProvider;
-import com.auth0.android.verification.JwtVerifier;
 import com.google.gson.Gson;
 import com.squareup.okhttp.HttpUrl;
 import com.squareup.okhttp.OkHttpClient;
@@ -107,7 +105,6 @@ public class AuthenticationAPIClient {
     private final Gson gson;
     private final RequestFactory factory;
     private final ErrorBuilder<AuthenticationException> authErrorBuilder;
-    private final JwtVerifier verifier;
 
 
     /**
@@ -139,18 +136,14 @@ public class AuthenticationAPIClient {
         this.client = clientFactory.createClient(auth0.isLoggingEnabled(), auth0.isTLS12Enforced());
         this.gson = gson;
         this.factory = factory;
+        JwtVerifier verifier = new JwtVerifier(new JwkProvider(auth0.getDomainUrl()));
+        verifier.setExpectedValues(auth0.getDomainUrl(), auth0.getClientId());
+        this.factory.setJwtVerifier(verifier);
         this.authErrorBuilder = new AuthenticationErrorBuilder();
-        this.verifier = new JwtVerifier(new JwkProvider(auth0.getDomainUrl()));
         final Telemetry telemetry = auth0.getTelemetry();
         if (telemetry != null) {
             factory.setClientInfo(telemetry.getValue());
         }
-    }
-
-    @NonNull
-    @VisibleForTesting
-    public JwtVerifier getTokenVerifier() {
-        return verifier;
     }
 
     public String getClientId() {
