+++ auth0/src/main/java/com/auth0/android/authentication/KeyProvider.java
@@ -1,0 +1,10 @@
+package com.auth0.android.authentication;
+
+import android.support.annotation.Nullable;
+
+import java.security.PublicKey;
+
+interface KeyProvider {
+
+    PublicKey getPublicKey(@Nullable String keyId) throws KeyProviderException;
+}
