+++ auth0/src/main/java/com/auth0/android/authentication/KeyProviderException.java
@@ -1,10 +1,10 @@
-package com.auth0.android.verification;
+package com.auth0.android.authentication;
 
 import com.auth0.android.Auth0Exception;
 
 class KeyProviderException extends Auth0Exception {
 
-    public KeyProviderException(String message, Throwable e) {
+    KeyProviderException(String message, Throwable e) {
         super(message, e);
     }
 }
