--- auth0/src/main/java/com/auth0/android/verification/KeyProvider.java
@@ -1,12 +1,0 @@
-package com.auth0.android.verification;
-
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-
-import com.auth0.android.callback.BaseCallback;
-
-import java.security.PublicKey;
-
-interface KeyProvider {
-    void getPublicKey(@Nullable String keyId, @NonNull BaseCallback<PublicKey, KeyProviderException> callback);
-}
