--- auth0/src/main/java/com/auth0/android/verification/TokenVerificationException.java
@@ -1,14 +1,0 @@
-package com.auth0.android.verification;
-
-import com.auth0.android.Auth0Exception;
-
-class TokenVerificationException extends Auth0Exception {
-
-    public TokenVerificationException(String message) {
-        super(message);
-    }
-
-    public TokenVerificationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
